kernel-build: moved compilation to own files
Moved
- function for llvm optimisation to `optimiser.py`,
- clang default compilation options to `compiler.py`,
so it can be reused by other snapshot builders.
